requirements.txt
pandas>=2.0.0
numpy>=1.23.0
plotly>=5.10.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
reportlab>=3.6.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
scikit-learn>=1.2.0
seaborn>=0.12.0
matplotlib>=3.5.0